refactor(users-pkg): package renaming #923
