+++ k2s/internal/core/users/users.go
@@ -15,12 +15,12 @@ import (
 	"github.com/siemens-healthineers/k2s/internal/core/node/ssh/keygen"
 	"github.com/siemens-healthineers/k2s/internal/core/users/acl"
 	"github.com/siemens-healthineers/k2s/internal/core/users/common"
+	"github.com/siemens-healthineers/k2s/internal/core/users/controlplane"
 	"github.com/siemens-healthineers/k2s/internal/core/users/fs"
 	"github.com/siemens-healthineers/k2s/internal/core/users/http"
 	"github.com/siemens-healthineers/k2s/internal/core/users/k8s"
 	"github.com/siemens-healthineers/k2s/internal/core/users/k8s/cluster"
 	"github.com/siemens-healthineers/k2s/internal/core/users/k8s/kubeconfig"
-	"github.com/siemens-healthineers/k2s/internal/core/users/nodes"
 	"github.com/siemens-healthineers/k2s/internal/core/users/winusers"
 )
 
@@ -74,7 +74,7 @@ func NewUsersManagement(cfg *config.Config, cmdExecutor common.CmdExecutor, user
 	aclExec := acl.NewAcl(cmdExecutor)
 	restClient := http.NewRestClient()
 	kubeconfigReader := kubeconfig.NewKubeconfigReader()
-	controlPlaneAccess := nodes.NewControlPlaneAccess(fileSystem, keygenExec, node.Exec, node.Copy, sshOptions, aclExec, cfg.Host.SshDir, controlePlaneCfg.IpAddress)
+	controlPlaneAccess := controlplane.NewControlPlaneAccess(fileSystem, keygenExec, node.Exec, node.Copy, sshOptions, aclExec, cfg.Host.SshDir, controlePlaneCfg.IpAddress)
 	clusterAccess := cluster.NewClusterAccess(restClient)
 	k8sAccess := k8s.NewK8sAccess(node.Exec, node.Copy, sshOptions, fileSystem, clusterAccess, kubeconfigWriterFactory, kubeconfigReader, cfg.Host.KubeConfigDir)
 	userAdder := NewWinUserAdder(controlPlaneAccess, k8sAccess, CreateK2sUserName)
